pyproject.toml
[project]
name = "agent_hello_world"
version = "0.1.0"
description = "一个最小可运行的中文 LangGraph Agent 基础项目。"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.5.4",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph-cli[inmem]>=0.3.1",
    "python-dotenv>=1.0.1",
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.6",
    "langchain-tavily",
    "tavily-python>=0.5.0",
    "mcp>=1.9.4",
    "aiohttp",
]

[project.scripts]
agent_hello_world-agent = "cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.6.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "cli",
    "configuration",
    "debug_trace",
    "graph",
    "model_factory",
    "prompts",
    "state",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP"]
